pyproject.toml
[tool.poetry]
name = "selfbot"
version = "0.1.0"
description = ""
authors = ["Ilia Arkania <iliaarkania13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py-self = "^2.0.0"
ollama = "^0.3.3"
colorama = "^0.4.6"
asyncio = "^3.4.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
